fix: change eventlistener on keyboard
