+++ src/style.css
@@ -23,6 +23,7 @@ body {
 }
 
 .wrapper .textarea {
+  position: relative;
   width: 100%;
   margin: 0 auto;
   max-width: 1400px;
@@ -36,16 +37,24 @@ body {
           box-sizing: border-box;
 }
 
+.wrapper .textarea__info {
+  position: absolute;
+  bottom: 5px;
+  left: 10px;
+  font-size: 1.5rem;
+  opacity: .4;
+}
+
 .wrapper .key {
   font-size: 2.1rem;
   margin: 0 auto;
   width: 100%;
-  height: 40vh;
+  height: 35vh;
   max-width: 1400px;
 }
 
 .wrapper .key .board {
-  height: 88%;
+  height: 100%;
   margin: 0 auto;
   display: -webkit-box;
   display: -ms-flexbox;
